+++ src/scvi/autotune/_experiment.py
@@ -2,8 +2,9 @@ from __future__ import annotations
 
 import inspect
 import logging
-from os.path import dirname, join
+from os.path import join
 from typing import TYPE_CHECKING
+from pathlib import Path
 
 import numpy as np
 import torch
@@ -296,7 +297,7 @@ class AutotuneExperiment:
         - ``"memory"``: amount of memory
 
         Passed into :func:`~ray.tune.with_resources`.
-    name
+    experiment_name
         Name of the experiment, used for logging purposes. Defaults to a unique ID.
     logging_dir
         Base directory to store experiment logs. Defaults to :attr:`~scvi.settings.logging_dir`.
@@ -320,6 +321,8 @@ class AutotuneExperiment:
     solver
         SVD solver to use during PCA. can help stability issues. Choose from: "arpack",
         "randomized" or "auto"
+    mudata_file_name
+        name of mudata file. can be a full path, but will not create folders
 
     Notes
     -----
@@ -342,7 +345,7 @@ class AutotuneExperiment:
         searcher: Literal["hyperopt", "random"] = "hyperopt",
         seed: int | None = None,
         resources: dict[Literal["cpu", "gpu", "memory"], float] | None = None,
-        name: str | None = None,
+        experiment_name: str | None = None,
         logging_dir: str | None = None,
         save_checkpoints: bool = False,
         scheduler_kwargs: dict | None = None,
@@ -352,24 +355,9 @@ class AutotuneExperiment:
         scib_indices_list: list | None = None,
         n_jobs: int = 1,
         solver: str = "arpack",
+        mudata_file_name: str = "mydata.h5mu",
     ) -> None:
         self.model_cls = model_cls
-        if type(data).__name__ == "MuData":
-            # save mudata on disk as it cant be pickled by ray
-            data.write_h5mu("mydata.h5mu")
-            self.is_mudata = True
-            # need to forcefully register it
-            data_manager = self.model_cls._get_most_recent_anndata_manager(data, required=True)
-            self._setup_method_name = data_manager._registry.get(
-                _SETUP_METHOD_NAME, "setup_anndata"
-            )
-            self._setup_method_args = data_manager._get_setup_method_args().get(
-                _SETUP_ARGS_KEY, {}
-            )
-            self.data = "mydata.h5mu"  # file will be read from the trainable folder upstream
-        else:
-            self.is_mudata = False
-            self.data = data
         self.metrics = metrics
         self.mode = mode
         self.search_space = search_space
@@ -380,7 +368,7 @@ class AutotuneExperiment:
         self.scheduler = scheduler
         self.searcher = searcher
         self.resources = resources
-        self.name = name
+        self.experiment_name = experiment_name
         self.logging_dir = logging_dir
         self.save_checkpoints = save_checkpoints
         self.scib_stage = scib_stage
@@ -388,6 +376,26 @@ class AutotuneExperiment:
         self.scib_indices_list = scib_indices_list
         self.n_jobs = n_jobs
         self.solver = solver
+        self.mudata_file_name = mudata_file_name
+
+        if type(data).__name__ == "MuData":
+            # save mudata on disk as it cant be pickled by ray
+            mudata_file_path = join(self._logging_dir, mudata_file_name)
+            Path(self._logging_dir).mkdir(parents=True, exist_ok=True)
+            data.write_h5mu(mudata_file_path)
+            self.is_mudata = True
+            # need to forcefully register it
+            data_manager = self.model_cls._get_most_recent_anndata_manager(data, required=True)
+            self._setup_method_name = data_manager._registry.get(
+                _SETUP_METHOD_NAME, "setup_anndata"
+            )
+            self._setup_method_args = data_manager._get_setup_method_args().get(
+                _SETUP_ARGS_KEY, {}
+            )
+            self.data = mudata_file_path  # file will be read from the trainable folder upstream
+        else:
+            self.is_mudata = False
+            self.data = data
 
     @property
     def id(self) -> str:
@@ -655,23 +663,23 @@ class AutotuneExperiment:
         self._resources = value or {}
 
     @property
-    def name(self) -> str:
+    def experiment_name(self) -> str:
         """Name of the experiment."""
-        if not hasattr(self, "_name"):
-            raise AttributeError("`name` not yet available.")
-        return self._name
+        if not hasattr(self, "_experiment_name"):
+            raise AttributeError("`experiment_name` not yet available.")
+        return self._experiment_name
 
-    @name.setter
-    def name(self, value: str | None) -> None:
-        if hasattr(self, "_name"):
-            raise AttributeError("Cannot reassign `name`")
+    @experiment_name.setter
+    def experiment_name(self, value: str | None) -> None:
+        if hasattr(self, "_experiment_name"):
+            raise AttributeError("Cannot reassign `experiment_name`")
         elif value is not None and not isinstance(value, str):
-            raise TypeError("`name` must be a string or `None`")
+            raise TypeError("`experiment_name` must be a string or `None`")
 
         if value is None:
             default = f"{self._model_cls.__name__.lower()}_"
             default += self.id
-        self._name = value or default
+        self._experiment_name = value or default
 
     @property
     def logging_dir(self) -> str:
@@ -688,7 +696,7 @@ class AutotuneExperiment:
             raise AttributeError("Cannot reassign `logging_dir`")
         elif value is not None and not isinstance(value, str):
             raise TypeError("`logging_dir` must be a string")
-        self._logging_dir = value or join(settings.logging_dir, self.name)
+        self._logging_dir = value or join(settings.logging_dir, self.experiment_name)
 
     @property
     def metrics_callback(self) -> Callback:
@@ -743,7 +751,7 @@ class AutotuneExperiment:
         self._result_grid = value
 
     def __repr__(self) -> str:
-        return f"Experiment {self.name}"
+        return f"Experiment {self.experiment_name}"
 
     def get_tuner(self) -> Tuner:
         """Configure a :class:`~ray.tune.Tuner` from this experiment."""
@@ -760,7 +768,7 @@ class AutotuneExperiment:
             num_samples=self.num_samples,
         )
         run_config = RunConfig(
-            name=self.name,
+            name=self.experiment_name,
             storage_path=self.logging_dir,
             log_to_file=True,
             verbose=1,
@@ -843,8 +851,7 @@ def _trainable(
         if experiment.is_mudata:
             import muon as mu
 
-            mudata_file_path = join(dirname(experiment.logging_dir), experiment.data)
-            adata_or_mdata = mu.read_h5mu(mudata_file_path)
+            adata_or_mdata = mu.read_h5mu(experiment.data)
         else:
             adata_or_mdata = experiment.data
         getattr(experiment.model_cls, experiment.setup_method_name)(